Create initial admin account from XXL_ env vars
The intended effect of this PR is to seed the initial admin account
credentials from the XXL_apiuser/XXL_apipass env vars when the
dockbix-xxl container is run with a fresh database endpoint.

As a automation engineer,
I want to control the initial Zabbix Super Admin account credentials,
so that deploy time customizations are not dependent on the insecure
default credentials.